% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{native_connect}
\alias{native_connect}
\title{Private connection primitive; does not construct a public DBI object.}
\usage{
native_connect(connection_string, config)
}
\description{
Private connection primitive; does not construct a public DBI object.
}
